Move layouts tests into new file
