+++ bokeh/models/component.py
@@ -7,7 +7,9 @@ from ..embed import notebook_div
 
 @abstract
 class Component(Model):
-    """ A base class for all embeddable models, i.e. plots and widgets. """
+    """ A base class for all embeddable models, i.e. plots, layouts and widgets.
+
+    """
 
     disabled = Bool(False, help="""
     Whether the widget will be disabled when rendered. If ``True``,
