+++ bokeh/models/widgets/__init__.py
@@ -8,7 +8,6 @@ from .dialogs import *
 from .groups import *
 from .icons import *
 from .inputs import *
-from .layouts import *
 from .markups import *
 from .panels import *
 from .tables import *
